FL-102 delete 시 serviceImpl에서 최근 id를 받도록 수정 (화면 연동 부분은 없음)

+++ flamingo-web/src/main/java/org/exem/flamingo/web/oozie/workflow/OozieWorkflowServiceImpl.java
@@ -131,7 +131,7 @@ public class OozieWorkflowServiceImpl implements OozieWorkflowService {
     oozieWorkflowRepository.updateWorkflow(param);
   }
 
-  public void deleteWorkflow(int id){
+  public void deleteWorkflow(long id){
     oozieWorkflowRepository.deleteWorkflow(id);
   }
 }
